requirements.txt
fastapi
uvicorn[standard]
supabase
python-dotenv
pandas
python-multipart
openpyxl
python-jose[cryptography]
passlib[bcrypt]
loguru